movies-django/requirements.txt
Django~=4.1.5
requests~=2.28.2
djangorestframework~=3.14.0
psycopg2-binary
django-filter
django-cors-headers
Markdown
py_eureka_client